feat: implement NavProgress component for top-of-page navigation indication
