+++ packages/ui/src/components/ui/nav-progress/index.ts
@@ -1,0 +1,1 @@
+export { default as NavProgress } from "./nav-progress.svelte";
